% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmc.R
\name{fetchPMC}
\alias{fetchPMC}
\title{Download one or more fulltext articles from PMC.}
\usage{
fetchPMC(pmcid)
}
\arguments{
\item{pmcid}{One or more PMC IDs

Returns a data frame with on row per PMC id, same order:
\itemize{
\item \code{pmc} The PMC ID of the article
\item \code{pmid} The PubMed ID of the article
\item \code{specificUse} The specific use the license applies to
\item \code{license} URL to the license the article is available under
\item \code{body} The XML body of the article
}

If the article is not available in PMC, most rows will be NA.
If PMC has the article, but does not allow XML download, the 'body' field
will be NA.

When the full text is available, the 'specificUse' will typically be 'textmining',
and the license some creative commons license, like https://creativecommons.org/licenses/by-nc-nd/4.0/

The body field contains the XML version of the article, inside \if{html}{\out{<body>}} tags.
It may be missing namespace definitions for e.g. mml: (MathML).}
}
\description{
Download one or more fulltext articles from PMC.
}
