% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{getArticles}
\alias{getArticles}
\title{Get articles from the DB}
\usage{
getArticles(from = "1000-01-01", to = "3000-01-01")
}
\arguments{
\item{from}{Start date (inclusive)}

\item{to}{End date (inclusive)}
}
\description{
Get articles from the DB
}
