% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getEndpoints}
\alias{getEndpoints}
\title{Get the endpoints defined in the settings file
Returns a list.}
\usage{
getEndpoints()
}
\description{
Get the endpoints defined in the settings file
Returns a list.
}
