% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{getEvals}
\alias{getEvals}
\title{Retreive evaluations (LLM answers)}
\usage{
getEvals(
  evaluator = NA,
  pmid = NA,
  before = "9999-12-31",
  after = "0001-01-01"
)
}
\arguments{
\item{evaluator}{Filter by evaluator}

\item{pmid}{Filter by Pubmed ID}

\item{before}{Filter by evaluation date}

\item{after}{Filter by evaluation date

Using all default values will fetch every single evaluation.
The evaluator and pmid fields accept multiple values.
The evaluators are ID numbers from \code{\link[=getOrRegisterEvaluator]{getOrRegisterEvaluator()}}.}
}
\description{
Retreive evaluations (LLM answers)
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Get answers to the 'Llama_example' question asked to phi4
# in 2025
ev = getOrRegisterEvaluator('phi4:latest', 'Llama_example')
answers = getEvals(evaluator=ev$id, after='2025-01-01')
\dontshow{\}) # examplesIf}
}
