% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{getOrRegisterEvaluator}
\alias{getOrRegisterEvaluator}
\title{Get info about an evaluator (a model + prompt pair)}
\usage{
getOrRegisterEvaluator(model, promptID)
}
\arguments{
\item{model}{The name of a model, e.g. "phi4:latest"}

\item{promptID}{The name of a prompt, from settings.xml}
}
\description{
Get info about an evaluator (a model + prompt pair)
}
