% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{getQueryByName}
\alias{getQueryByName}
\title{Get a pubmed query by name from the DB}
\usage{
getQueryByName(name)
}
\arguments{
\item{name}{Name of a query (from settings.xml)}
}
\description{
Get a pubmed query by name from the DB
}
