% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{getTypes}
\alias{getTypes}
\title{Get all known article types and counts}
\usage{
getTypes()
}
\description{
Returns a data frame with 'name' and 'count',
where the count is the number of articles in the entire DB with this type.
One article can have multiple types.
}
