% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{grabJSON}
\alias{grabJSON}
\title{Grab something JSON-like from a string}
\usage{
grabJSON(text)
}
\arguments{
\item{text}{The text to search

Returns anything from the first \{ to the last \}, inclusive.}
}
\description{
Grab something JSON-like from a string
}
