% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{insertEvals}
\alias{insertEvals}
\title{Store an answer from an evaluator for a given article}
\usage{
insertEvals(evaluatorID, pmid, answer)
}
\arguments{
\item{evaluatorID}{Evaluator, typically from getOrRegisterEvaluator}

\item{pmid}{One or more pubmed id numbers}

\item{answer}{One or more answers (arbitrary string, doesn't have to be valid JSON)}
}
\description{
Store an answer from an evaluator for a given article
}
