% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{jsonToTable}
\alias{jsonToTable}
\title{Explode a vector of JSON strings to a table}
\usage{
jsonToTable(raw_json, NA_string = "NA")
}
\arguments{
\item{raw_json}{A vector of strings containing unparsed JSON text}

\item{NA_string}{The value used to fill gaps from missing values (default "NA")

This is a "best effort" attempt: It fills gaps and missing values
and attempts to return a data.frame with as may rows as the input,
and with one column for each unique key seen in the JSON, padded out
with the NA_string value where needed.}
}
\description{
Explode a vector of JSON strings to a table
}
