% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{lookupPmidPmc}
\alias{lookupPmidPmc}
\title{Get known PMID to PMC conversions from the DB}
\usage{
lookupPmidPmc(pmid)
}
\arguments{
\item{pmid}{one or more PMIDs

Returns a data.frame with PMID and PMC (or NA)}
}
\description{
Get known PMID to PMC conversions from the DB
}
