% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmc.R
\name{pmid2pmcid}
\alias{pmid2pmcid}
\title{Convert Pubmed to PMC IDs}
\usage{
pmid2pmcid(pmid)
}
\arguments{
\item{pmid}{One or mode PMIDs to convert

Returns a vector of the same length as the input,
with corresponding PMC IDs or NA.}
}
\description{
Convert Pubmed to PMC IDs
}
