% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed.R
\name{searchPubmed}
\alias{searchPubmed}
\title{Performs a named pubmed query, stores the results}
\usage{
searchPubmed(queryname)
}
\arguments{
\item{queryname}{Name of the query to run

Queries and their names are defined in the settings file, see publlamaInit().}
}
\description{
Performs a named pubmed query, stores the results
}
