% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{strip}
\alias{strip}
\title{Remove leading and trailing whitespace from a string}
\usage{
strip(str)
}
\arguments{
\item{str}{The string to trim}
}
\description{
Remove leading and trailing whitespace from a string
}
